requirements.txt
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.0.20
python-dotenv>=1.0.0
streamlit
